+++ src/canisters/frontend/ic-certified-assets/src/asset_certification/types/rc_bytes.rs
